refactor: code review feedback
