+++ src/ape_pm/compiler.py
@@ -1,6 +1,6 @@
 import json
 from pathlib import Path
-from typing import List
+from typing import List, Set
 
 from ape.api.compiler import CompilerAPI
 from ape.types import ContractType, PackageManifest
@@ -11,11 +11,10 @@ class PackageCompiler(CompilerAPI):
     def name(self) -> str:
         return "ethpm"
 
-    @property
-    def versions(self) -> List[str]:
+    def get_versions(self, all_paths: List[Path]) -> Set[str]:
         # NOTE: This bypasses the serialization of this compiler into the package manifest's
         #       `compilers` field. You should not do this with a real compiler plugin.
-        return []
+        return set()
 
     def compile(self, filepaths: List[Path]) -> List[ContractType]:
         contract_types: List[ContractType] = []
